pyproject.toml
[project]
name = "bajaj-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "fastapi>=0.122.0",
    "imageio[pyav]>=2.37.2",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
